extract renderer-util from renderer
